+++ kdl_wagtail/core/models.py
@@ -1,38 +1,15 @@
-from django.conf import settings
-from django.core.paginator import EmptyPage, PageNotAnInteger, Paginator
 from django.db import models
 from wagtail.admin.edit_handlers import FieldPanel, StreamFieldPanel
 from wagtail.api import APIField
+from wagtail.contrib.settings.models import BaseSetting, register_setting
 from wagtail.core.fields import RichTextField, StreamField
 from wagtail.core.models import Page
 from wagtail.images.edit_handlers import ImageChooserPanel
 from wagtail.images.models import Image
 from wagtail.search import index
-from wagtail.snippets.models import register_snippet
 
 from .blocks import BaseStreamBlock
-
-
-@register_snippet
-class FooterText(models.Model):
-    """
-    This provides editable text for the site footer. It is made accessible in
-    the templates via the template tag `get_footer_text` defined in
-    `kdl_wagtail/templatetags/kdl_wagtail_core_tags.py`.
-    It uses the `@register_snippet` decorator to allow it to be accessible
-    via the Snippets UI (e.g. /admin/snippets/kdl_wagtail_core/footertext/)
-    """
-    body = RichTextField()
-
-    panels = [
-        FieldPanel('body'),
-    ]
-
-    class Meta:
-        verbose_name_plural = 'Footer Text'
-
-    def __str__(self):
-        return 'Footer text'
+from .utils import paginate
 
 
 class BasePage(Page):
@@ -79,25 +56,20 @@ class BaseIndexPage(BasePage):
     def get_context(self, request):
         context = super().get_context(request)
 
-        children = self.paginate(request, self.children())
-
+        children = self._paginate(request)
         context['children'] = children
 
         return context
 
-    def paginate(self, request, *args):
-        page = request.GET.get('page')
-        paginator = Paginator(
-            self.children(), settings.KDL_WAGTAIL_ITEMS_PER_PAGE)
+    def _paginate(self, request):
+        return paginate(self.children(), request.GET.get('page'))
 
-        try:
-            pages = paginator.page(page)
-        except PageNotAnInteger:
-            pages = paginator.page(1)
-        except EmptyPage:
-            pages = paginator.page(paginator.num_pages)
 
-        return pages
+class IndexPage(BaseIndexPage):
+    """
+    A direct implementation of `BaseIndexPage`.
+    """
+    pass
 
 
 class RichTextPage(BasePage):
@@ -137,3 +109,13 @@ class StreamPage(BasePage):
     search_fields = BasePage.search_fields + [
         index.SearchField('body')
     ]
+
+
+@register_setting
+class AnalyticsSettings(BaseSetting):
+    analytics_id = models.CharField(max_length=255)
+
+
+@register_setting
+class FooterSettings(BaseSetting):
+    body = RichTextField()