+++ HISTORY.rst
@@ -3,6 +3,15 @@
 History
 -------
 
+0.5.0 (2019-02-05)
+* Wagtail settings for footer text and analytics settings
+* `kdl_wagtail_core` utils:
+    * `paginate`
+* `kdl_wagtail_core` template tags:
+    * `get_page_children`
+    * `get_analytics_id`
+
+
 0.1.0 (2019-01-22)
 ++++++++++++++++++